refactor(plugin-sdk): share planned migration target resolution
