+++ extensions/migrate-hermes/targets.ts
@@ -1,31 +1,5 @@
-// Migrate Hermes plugin module implements targets behavior.
-import path from "node:path";
-import {
-  resolveAgentConfig,
-  resolveAgentWorkspaceDir,
-  resolveDefaultAgentId,
-} from "openclaw/plugin-sdk/agent-runtime";
-import type { MigrationProviderContext } from "openclaw/plugin-sdk/plugin-entry";
-import { resolveHomePath } from "./helpers.js";
-
-export type PlannedTargets = {
-  workspaceDir: string;
-  stateDir: string;
-  agentDir: string;
-};
-
-export function resolveTargets(ctx: MigrationProviderContext): PlannedTargets {
-  const cfg = ctx.config;
-  const agentId = resolveDefaultAgentId(cfg);
-  const workspaceDir = resolveAgentWorkspaceDir(cfg, agentId);
-  const configuredAgentDir = resolveAgentConfig(cfg, agentId)?.agentDir?.trim();
-  const agentDir =
-    ctx.runtime?.agent?.resolveAgentDir(cfg, agentId) ??
-    (configuredAgentDir ? resolveHomePath(configuredAgentDir) : undefined) ??
-    path.join(ctx.stateDir, "agents", agentId, "agent");
-  return {
-    workspaceDir,
-    stateDir: ctx.stateDir,
-    agentDir,
-  };
-}
+// Migrate Hermes plugin re-exports the shared migration target resolution.
+export {
+  resolvePlannedMigrationTargets as resolveTargets,
+  type PlannedMigrationTargets as PlannedTargets,
+} from "openclaw/plugin-sdk/migration-runtime";
