+++ src/plugin-sdk/migration-runtime.ts
@@ -3,7 +3,10 @@
 import crypto from "node:crypto";
 import fs from "node:fs/promises";
 import path from "node:path";
+import { resolveAgentConfig } from "../agents/agent-scope-config.js";
+import { resolveAgentWorkspaceDir, resolveDefaultAgentId } from "../agents/agent-scope.js";
 import { pathExists } from "../infra/fs-safe.js";
+import { resolveHomeRelativePath } from "../infra/home-dir.js";
 import type {
   MigrationApplyResult,
   MigrationItem,
@@ -18,6 +21,37 @@ import {
 } from "./migration.js";
 
 export type { MigrationApplyResult, MigrationItem } from "../plugins/types.js";
+
+/** Directories a migration provider writes imported agent data into. */
+export type PlannedMigrationTargets = {
+  workspaceDir: string;
+  stateDir: string;
+  agentDir: string;
+};
+
+/**
+ * Resolves the default agent's workspace/state/agent directories for a
+ * migration run. Prefers the runtime resolver, then a configured agentDir
+ * (home-relative paths honor the effective-home resolution used by the rest
+ * of the product), then the canonical state-dir layout.
+ */
+export function resolvePlannedMigrationTargets(
+  ctx: MigrationProviderContext,
+): PlannedMigrationTargets {
+  const cfg = ctx.config;
+  const agentId = resolveDefaultAgentId(cfg);
+  const workspaceDir = resolveAgentWorkspaceDir(cfg, agentId);
+  const configuredAgentDir = resolveAgentConfig(cfg, agentId)?.agentDir?.trim();
+  const agentDir =
+    ctx.runtime?.agent?.resolveAgentDir(cfg, agentId) ??
+    (configuredAgentDir ? resolveHomeRelativePath(configuredAgentDir) : undefined) ??
+    path.join(ctx.stateDir, "agents", agentId, "agent");
+  return {
+    workspaceDir,
+    stateDir: ctx.stateDir,
+    agentDir,
+  };
+}
 
 /** Wrap migration runtime config access with a cached mutable snapshot during apply. */
 export function withCachedMigrationConfigRuntime(
